Renamed classes, ident horizontal layout with 4 spaces
