+++ CardsCollectionViewHorizontalLayout.swift
@@ -1,5 +1,5 @@
 //
-//  CardsCollectionViewVerticleLayout.swift
+//  CardsCollectionViewHorizontalLayout.swift
 //  CardsLayout
 //
 //  Created by Filipp Fediakov on 18.08.17.
@@ -8,57 +8,59 @@
 
 import UIKit
 
-open class CardsCollectionViewVerticleLayout: UICollectionViewLayout {
-    
+open class CardsCollectionViewHorizontalLayout: UICollectionViewLayout {
+
     // MARK: - Layout configuration
-    
+
     public var itemSize: CGSize = CGSize(width: 200, height: 300) {
-      didSet{
-        invalidateLayout()
-      }
-    }
-    
-    public var spacing: CGPoint = CGPoint(x: 10.0, y: 20.0) {
-        didSet{
+        didSet {
             invalidateLayout()
         }
     }
-    
+
+    public var spacing: CGFloat = 10.0 {
+        didSet {
+            invalidateLayout()
+        }
+    }
+
     public var maximumVisibleItems: Int = 4 {
-        didSet{
+        didSet {
             invalidateLayout()
         }
     }
-    
+
     // MARK: UICollectionViewLayout
-    
+
     override open var collectionView: UICollectionView {
         return super.collectionView!
     }
-    
+
     override open var collectionViewContentSize: CGSize {
         let itemsCount = CGFloat(collectionView.numberOfItems(inSection: 0))
-        return CGSize(width: collectionView.bounds.width,
-                      height: collectionView.bounds.height * itemsCount)
+        return CGSize(width: collectionView.bounds.width * itemsCount,
+                      height: collectionView.bounds.height)
     }
-    
+
     override open func prepare() {
         super.prepare()
         assert(collectionView.numberOfSections == 1, "Multiple sections aren't supported!")
     }
-    
+
     override open func layoutAttributesForElements(in rect: CGRect) -> [UICollectionViewLayoutAttributes]? {
         let totalItemsCount = collectionView.numberOfItems(inSection: 0)
 
-        let minVisibleIndex = max(Int(collectionView.contentOffset.y) / Int(collectionView.bounds.height), 0)
+        let minVisibleIndex = max(Int(collectionView.contentOffset.x) / Int(collectionView.bounds.width), 0)
         let maxVisibleIndex = min(minVisibleIndex + maximumVisibleItems, totalItemsCount)
-        
+
         let contentCenterX = collectionView.contentOffset.x + (collectionView.bounds.width / 2.0)
-        let deltaOffset = Int(collectionView.contentOffset.y) % Int(collectionView.bounds.height)
-        let percentageDeltaOffset = CGFloat(deltaOffset) / collectionView.bounds.height
+
+        let deltaOffset = Int(collectionView.contentOffset.x) % Int(collectionView.bounds.width)
+
+        let percentageDeltaOffset = CGFloat(deltaOffset) / collectionView.bounds.width
+
         let visibleIndices = minVisibleIndex..<maxVisibleIndex
-        print("Called layoutAttributesForElements rect = \(rect), delta = \(deltaOffset), percentage = \(percentageDeltaOffset)%")
-        
+
         let attributes: [UICollectionViewLayoutAttributes] = visibleIndices.map { index in
             let indexPath = IndexPath(item: index, section: 0)
             return computeLayoutAttributesForItem(indexPath: indexPath,
@@ -67,24 +69,22 @@ open class CardsCollectionViewVerticleLayout: UICollectionViewLayout {
                                                   deltaOffset: CGFloat(deltaOffset),
                                                   percentageDeltaOffset: percentageDeltaOffset)
         }
-        
+
         return attributes
     }
-    
-    override open func layoutAttributesForItem(at indexPath: IndexPath) -> UICollectionViewLayoutAttributes? {
 
+    override open func layoutAttributesForItem(at indexPath: IndexPath) -> UICollectionViewLayoutAttributes? {
         let contentCenterX = collectionView.contentOffset.x + (collectionView.bounds.width / 2.0)
         let minVisibleIndex = Int(collectionView.contentOffset.x) / Int(collectionView.bounds.width)
-        let deltaOffset = Int(collectionView.contentOffset.y) % Int(collectionView.bounds.height)
-        let percentageDeltaOffset = CGFloat(deltaOffset) / collectionView.bounds.height
-        print("Called layoutAttributesForItem IP = \(indexPath), delta = \(deltaOffset), percentage = \(percentageDeltaOffset)%")
+        let deltaOffset = Int(collectionView.contentOffset.x) % Int(collectionView.bounds.width)
+        let percentageDeltaOffset = CGFloat(deltaOffset) / collectionView.bounds.width
         return computeLayoutAttributesForItem(indexPath: indexPath,
                                               minVisibleIndex: minVisibleIndex,
                                               contentCenterX: contentCenterX,
                                               deltaOffset: CGFloat(deltaOffset),
                                               percentageDeltaOffset: percentageDeltaOffset)
     }
-    
+
     override open func shouldInvalidateLayout(forBoundsChange newBounds: CGRect) -> Bool {
         return true
     }
@@ -92,8 +92,9 @@ open class CardsCollectionViewVerticleLayout: UICollectionViewLayout {
 
 
 // MARK: - Layout computations
-fileprivate extension CardsCollectionViewVerticleLayout {
-    
+
+fileprivate extension CardsCollectionViewHorizontalLayout {
+
     private func scale(at index: Int) -> CGFloat {
         let translatedCoefficient = CGFloat(index) - CGFloat(self.maximumVisibleItems) / 2
         return CGFloat(pow(0.95, translatedCoefficient))
@@ -109,7 +110,7 @@ fileprivate extension CardsCollectionViewVerticleLayout {
         }
         return CGAffineTransform(scaleX: rawScale, y: rawScale)
     }
-    
+
     fileprivate func computeLayoutAttributesForItem(indexPath: IndexPath,
                                                     minVisibleIndex: Int,
                                                     contentCenterX: CGFloat,
@@ -117,25 +118,23 @@ fileprivate extension CardsCollectionViewVerticleLayout {
                                                     percentageDeltaOffset: CGFloat) -> UICollectionViewLayoutAttributes {
         let attributes = UICollectionViewLayoutAttributes(forCellWith:indexPath)
         let visibleIndex = indexPath.row - minVisibleIndex
-       
-        attributes.size = self.itemSize
-        
+        attributes.size = itemSize
         let midY = self.collectionView.bounds.midY
-        let midX = self.collectionView.bounds.midX
-        attributes.center = CGPoint(x: midX + spacing.x * CGFloat(visibleIndex),
-                                    y: midY + spacing.y * CGFloat(visibleIndex))
+        attributes.center = CGPoint(x: contentCenterX + spacing * CGFloat(visibleIndex),
+                                    y: midY + spacing * CGFloat(visibleIndex))
         attributes.zIndex = maximumVisibleItems - visibleIndex
-        
+
         attributes.transform = transform(atCurrentVisibleIndex: visibleIndex,
                                          percentageOffset: percentageDeltaOffset)
         switch visibleIndex {
         case 0:
-            attributes.center.y -= deltaOffset
+            attributes.center.x -= deltaOffset
             break
         case 1..<maximumVisibleItems:
-            attributes.center.x -= spacing.x * percentageDeltaOffset
-            attributes.center.y -= spacing.y * percentageDeltaOffset
-            
+            attributes.center.x -= spacing * percentageDeltaOffset
+            attributes.center.y -= spacing * percentageDeltaOffset
+
+
             if visibleIndex == maximumVisibleItems - 1 {
                 attributes.alpha = percentageDeltaOffset
             }
@@ -147,3 +146,4 @@ fileprivate extension CardsCollectionViewVerticleLayout {
         return attributes
     }
 }
+
